+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # What's New?
 
-## 1.20
+## 1.20.x
+
+Improvements:
+
+- Add notification suggesting users to uninstall twxs.cmake now that we have built-in Language Services. [#4288](https://github.com/microsoft/vscode-cmake-tools/issues/4288)
+
+Bug Fixes:
+
+- Fix regression where we weren't properly expanding copyCompileCommands path. [#4294](https://github.com/microsoft/vscode-cmake-tools/issues/4294)
+
+## 1.20.52
 
 Features: